+++ src/test/java/com/market/board/infrastructure/image/ImageFakeUploader.java
@@ -1,7 +1,7 @@
 package com.market.board.infrastructure.image;
 
+import com.market.board.application.ImageUploader;
 import com.market.board.domain.image.Image;
-import com.market.board.domain.image.ImageUploader;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.util.List;
